Rename package tools to dto - (it doesn't contain tools anymore)
